Move database call to different file
