+++ composeApp/src/commonMain/kotlin/com/daily/vitals/TestDatabaseCall.kt
@@ -1,0 +1,41 @@
+package com.daily.vitals
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.Button
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.rememberCoroutineScope
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import dev.gitlive.firebase.Firebase
+import dev.gitlive.firebase.database.database
+import kotlinx.coroutines.launch
+
+@Composable
+fun TestDatabaseCall() {
+    val scope = rememberCoroutineScope()
+    Column(
+        modifier = Modifier.fillMaxSize(),
+        verticalArrangement = Arrangement.Center,
+        horizontalAlignment = Alignment.CenterHorizontally
+    ) {
+        Button(
+            onClick = {
+                scope.launch {
+                    writeTest()
+                }
+            },
+            content = {
+                Text("click here")
+            }
+        )
+    }
+
+}
+
+suspend fun writeTest() {
+    val db = Firebase.database
+    db.reference().child("test-iOS").setValue("Hello, world!")
+}
